move data/prepare.py into ann_benchmarks/datasets.py, create entry point"
